+++ src/app/(developers)/developers/frame/page.tsx
@@ -9,7 +9,7 @@ import urlcat from 'urlcat';
 import { Headline } from '@/app/(settings)/components/Headline.js';
 import { ClickableButton } from '@/components/ClickableButton.js';
 import { Frame as FrameUI } from '@/components/Frame/index.js';
-import { URL_REGEX } from '@/constants/regex.js';
+import { URL_REGEX } from '@/constants/regexp.js';
 import { classNames } from '@/helpers/classNames.js';
 import { fetchJSON } from '@/helpers/fetchJSON.js';
