+++ src/components/Markup/ArticleMarkup.tsx
@@ -11,7 +11,7 @@ import stripMarkdown from 'strip-markdown';
 import { Code } from '@/components/Code.js';
 import { MarkupLink } from '@/components/Markup/MarkupLink/index.js';
 import { ImageAsset } from '@/components/Posts/ImageAsset.js';
-import { BIO_TWITTER_PROFILE_REGEX, URL_REGEX } from '@/constants/regex.js';
+import { BIO_TWITTER_PROFILE_REGEX, URL_REGEX } from '@/constants/regexp.js';
 import { classNames } from '@/helpers/classNames.js';
 import { trimify } from '@/helpers/trimify.js';
 import { PreviewImageModalRef } from '@/modals/controls.js';
