+++ src/constants/error.ts
@@ -1,0 +1,5 @@
+export class AbortError extends Error {
+    constructor() {
+        super('Aborted');
+    }
+}
