+++ src/helpers/formatFarcasterPostFromFirefly.ts
@@ -1,7 +1,7 @@
 import { compact, first, last, uniqBy } from 'lodash-es';
 
 import { Source } from '@/constants/enum.js';
-import { URL_REGEX } from '@/constants/regex.js';
+import { URL_REGEX } from '@/constants/regexp.js';
 import { fixUrlProtocol } from '@/helpers/fixUrlProtocol.js';
 import { formatChannelFromFirefly } from '@/helpers/formatFarcasterChannelFromFirefly.js';
 import { formatFarcasterProfileFromFirefly } from '@/helpers/formatFarcasterProfileFromFirefly.js';
