+++ src/mask/hooks/usePostInfo.ts
@@ -15,7 +15,7 @@ import { useMemo } from 'react';
 import urlcat from 'urlcat';
 
 import { SITE_HOSTNAME, SITE_URL } from '@/constants/index.js';
-import { URL_REGEX } from '@/constants/regex.js';
+import { URL_REGEX } from '@/constants/regexp.js';
 import { getPostUrl } from '@/helpers/getPostUrl.js';
 import type { Post } from '@/providers/types/SocialMedia.js';
