+++ src/providers/firefly/Session.ts
@@ -39,6 +39,7 @@ export class FireflySession extends BaseSession implements Session {
             }
             case SessionType.Farcaster: {
                 if (FarcasterSession.isCustodyWallet(session)) throw new Error('Not allowed');
+
                 const isGrantByPermission = FarcasterSession.isGrantByPermission(session);
                 const isRelayService = FarcasterSession.isRelayService(session);
